requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.1.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
config==0.4.2
heroku==0.1.4
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
pathspec==0.8.0
Pillow==7.2.0
pylint==2.6.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-dateutil==1.5
regex==2020.7.14
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
